ejericio 1 en proceso

+++ app/src/main/java/com/example/android_ejercicios_xml/domain/models/Cronograma.kt
@@ -7,5 +7,5 @@ data class Cronograma(
     val seguros: Double,
     val subvencion: Double,
     val cuota: Double,
-    val saldo: Double
+    val saldo: Float
 )
